Some visual changes and reconstruct command arguments
